mostly for closing bug #88712 Package-Manager: portage-2.0.51.19

+++ net-dns/bind/bind-9.2.5-r1.ebuild
@@ -1,20 +1,28 @@
 # Copyright 1999-2005 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/net-dns/bind/bind-9.2.2-r1.ebuild,v 1.11 2005/01/01 16:19:26 weeve Exp $
+# $Header: /var/cvsroot/gentoo-x86/net-dns/bind/bind-9.2.5-r1.ebuild,v 1.1 2005/04/12 11:01:19 voxus Exp $
 
-IUSE="doc ipv6 selinux ssl"
+inherit eutils gnuconfig libtool
+
+IUSE="ssl ipv6 doc dlz postgres berkdb bind-mysql mysql odbc ldap selinux \
+	idn caps threads"
 
 DESCRIPTION="BIND - Berkeley Internet Name Domain - Name Server"
-SRC_URI="ftp://ftp.isc.org/isc/bind9/${PV}/${P}.tar.gz"
+SRC_URI="ftp://ftp.isc.org/isc/bind9/${PV}/${P}.tar.gz
+	dlz? ( http://dev.gentoo.org/~voxus/dlz/dlz-${PV}.patch.bz2 )"
 HOMEPAGE="http://www.isc.org/products/BIND/bind9.html"
 
-KEYWORDS="x86 ppc ~sparc ~alpha"
+KEYWORDS="~x86 ~ppc ~sparc ~alpha ~hppa ~amd64 ~ppc64 ~mips"
+
 LICENSE="as-is"
 SLOT="0"
 
 DEPEND="sys-apps/groff
+	sys-devel/autoconf
 	>=sys-apps/sed-4
-	ssl? ( >=dev-libs/openssl-0.9.6g )"
+	ssl? ( >=dev-libs/openssl-0.9.6g )
+	mysql? ( >=dev-db/mysql-4 )
+	ldap? ( net-nds/openldap )"
 
 RDEPEND="${DEPEND}
 	selinux? ( sec-policy/selinux-bind )"
@@ -23,27 +31,72 @@ src_unpack() {
 	unpack ${A} && cd ${S}
 
 	# Adjusting PATHs in manpages
-	for i in `echo bin/{named/named.8,check/named-checkconf.8,nsupdate/nsupdate.8,rndc/rndc.8}`; do
+	for i in `echo bin/{named/named.8,check/named-checkconf.8,rndc/rndc.8}`; do
 		sed -i -e 's:/etc/named.conf:/etc/bind/named.conf:g' \
 		       -e 's:/etc/rndc.conf:/etc/bind/rndc.conf:g' \
 		       -e 's:/etc/rndc.key:/etc/bind/rndc.key:g' \
 		       ${i}
 	done
+
+	if use dlz; then
+		epatch ${DISTDIR}/dlz-${PV}.patch.bz2
+		epatch ${FILESDIR}/${P}-berkdb_fix.patch
+	fi
+
+	if use bind-mysql; then
+		if use dlz; then
+			epatch ${FILESDIR}/${P}-dlz-mysql.patch
+		else
+			epatch ${FILESDIR}/${P}-mysql.patch
+		fi
+	fi
+
+	if use idn; then
+		epatch ${S}/contrib/idn/idnkit-1.0-src/patch/bind9/${P}-patch
+	fi
+
+	cp ${FILESDIR}/named.rc6 ${T}
+	cd ${T} && epatch ${FILESDIR}/named.rc6-pid_fix
+
+	gnuconfig_update
+
+	# it should be installed by bind-tools
+	sed "s:nsupdate ::g" ${S}/bin/Makefile.in > ${T}/Makefile
+	mv ${T}/Makefile ${S}/bin/Makefile.in
+
+	cd ${S}
+	WANT_AUTOCONF=2.5 autoconf || die "autoconf failed"
 }
 
 src_compile() {
 	local myconf=""
 
 	use ssl && myconf="${myconf} --with-openssl"
-	use ipv6 && myconf="${myconf} --enable-ipv6" || myconf="${myconf} --enable-ipv6=no"
+	use dlz && {
+		myconf="${myconf} --with-dlz-filesystem --with-dlz-stub"
+		use postgres && myconf="${myconf} --with-dlz-postgres"
+		use mysql && myconf="${myconf} --with-dlz-mysql"
+		use berkdb && myconf="${myconf} --with-dlz-bdb"
+		use ldap  && myconf="${myconf} --with-dlz-ldap"
+		use odbc  && myconf="${myconf} --with-dlz-odbc"
+	}
+	use caps || myconf="${myconf} --disable-linux-caps"
 
-	econf 	--sysconfdir=/etc/bind \
+	econf \
+		--sysconfdir=/etc/bind \
 		--localstatedir=/var \
-		--enable-threads \
+		`use_enable threads` \
+		`use_enable ipv6` \
 		--with-libtool \
 		${myconf} || die "econf failed"
 
-	MAKEOPTS="${MAKEOPTS} -j1" emake || die "failed to compile bind"
+	emake -j1 || die "failed to compile bind"
+
+	if use idn; then
+		cd ${S}/contrib/idn/idnkit-1.0-src
+		econf || die "idn econf failed"
+		emake || die "idn emake failed"
+	fi
 }
 
 src_install() {
@@ -53,7 +106,7 @@ src_install() {
 
 	use doc && {
 		docinto misc ; dodoc doc/misc/*
-		docinto html ; dodoc doc/arm/*
+		docinto html ; dohtml doc/arm/*
 		docinto	draft ; dodoc doc/draft/*
 		docinto rfc ; dodoc doc/rfc/*
 		docinto contrib ; dodoc contrib/named-bootconf/named-bootconf.sh \
@@ -75,17 +128,22 @@ src_install() {
 	insinto /var/bind ; doins ${FILESDIR}/named.ca
 	insinto /var/bind/pri ; doins ${FILESDIR}/{127,localhost}.zone
 
-	exeinto /etc/init.d ; newexe ${FILESDIR}/named.rc6 named
+	exeinto /etc/init.d ; newexe ${T}/named.rc6 named
 	insinto /etc/conf.d ; newins ${FILESDIR}/named.confd named
 
 	dosym ../../var/bind/named.ca /var/bind/root.cache
 	dosym ../../var/bind/pri /etc/bind/pri
 	dosym ../../var/bind/sec /etc/bind/sec
-}
 
-pkg_preinst() {
+	if use idn; then
+		cd ${S}/contrib/idn/idnkit-1.0-src
+		einstall || die "failed to install idn kit"
+		docinto idn
+		dodoc ChangeLog INSTALL{,.ja} README{,.ja} NEWS
+	fi
+
 	# Let's get rid of those tools and their manpages since they're provided by bind-tools
-	rm -f ${D}/usr/share/man/man1/{dig.1.gz,host.1.gz}
+	rm -f ${D}/usr/share/man/man1/{dig.1,host.1,nslookup.1}
 	rm -f ${D}/usr/bin/{dig,host,nslookup}
 }
 
@@ -110,11 +168,36 @@ pkg_postinst() {
 	einfo "\`ebuild /var/db/pkg/${CATEGORY}/${PF}/${PF}.ebuild config\`"
 	einfo "Before running the above command you might want to change the chroot"
 	einfo "dir in /etc/conf.d/named. Otherwise /chroot/dns will be used."
-	einfo
+	echo
+	einfo "Recently verisign added a wildcard A record to the .COM and .NET TLD"
+	einfo "zones making all .com and .net domains appear to be registered"
+	einfo "This causes many problems such as breaking important anti-spam checks"
+	einfo "which verify source domains exist. ISC released a patch for BIND which"
+	einfo "adds 'delegation-only' zones to allow admins to return the .com and .net"
+	einfo "domain resolution to their normal function."
+	echo
+	einfo "There is no need to create a com or net data file. Just the"
+	einfo "entries to the named.conf file is enough."
+	echo
+	einfo "	zone "com" IN { type delegation-only; };"
+	einfo "	zone "net" IN { type delegation-only; };"
+
+	if use dlz && use mysql; then
+		echo
+		ewarn ""
+		einfo "MySQL uses thread local storage in its C api. Thus MySQL"
+		einfo "requires that each thread of an application execute a MySQL"
+		einfo "\"thread initialization\" to setup the thread local storage."
+		einfo "This is impossible to do safely while staying within the DLZ"
+		einfo "driver API. This is a limitation caused by MySQL, and not the"
+		einfo "DLZ API."
+		ewarn "Because of this BIND MUST only run with a single thread when"
+		ewarn "using the MySQL driver."
+		echo
+	fi
 }
 
 pkg_config() {
-
 	CHROOT=`sed -n 's/^[[:blank:]]\?CHROOT="\([^"]\+\)"/\1/p' /etc/conf.d/named 2>/dev/null`
 	EXISTS="no"
 
@@ -154,3 +237,4 @@ pkg_config() {
 		einfo
 	fi
 }
+
